+++ knock/src/components/QuestionList/QuestionCardList.module.scss
@@ -3,21 +3,15 @@
   justify-content: center;
   margin: 0 2rem;
   gap: 1.25rem;
+  grid-template: repeat(2, 1fr) / repeat(4, minmax(180px, 220px));
 
-  @media (width <= 430px) {
+  @media (width < 876px) {
+    grid-template: repeat(2, 1fr) / repeat(3, minmax(180px, 220px));
+  }
+  @media (width < 668px) {
+    grid-template: repeat(3, 1fr) / repeat(2, 1fr);
+    gap: 1rem;
     margin: 0 1.5rem;
-  }
-
-  &--large {
-    grid-template: repeat(2, 1fr) / repeat(4, minmax(186px, 220px));
-  }
-
-  &--small {
-    grid-template: repeat(2, 1fr) / repeat(3, minmax(186px, 220px));
-    @media (width <= 430px) {
-      grid-template: repeat(3, 1fr) / repeat(2, 1fr);
-      gap: 1rem;
-    }
   }
 
   & a {
@@ -25,7 +19,7 @@
   }
   & li {
     cursor: pointer;
-    @media (width <= 430px) {
+    @media (width < 668px) {
       width: 9.71875rem;
     }
   }
