+++ knock/src/components/QuestionList/QuestionListPagination.tsx
@@ -6,6 +6,7 @@ import styles from './QuestionListPagination.module.scss';
 import { SubjectDetailResponse } from '../../core/types/api/Response';
 import { SubjectListParams } from '../../core/types/api/Request';
 import { getSubjectList } from '../../lib/api/Subject';
+import useResize from '../../lib/hooks/useResize';
 
 interface QuestionListPaginationProps {
   order: string;
@@ -15,9 +16,8 @@ const QuestionListPagination = ({ order }: QuestionListPaginationProps) => {
   const [questions, setQuestions] = useState<SubjectDetailResponse[]>([]);
   const [pageIndexes, setPageIndexes] = useState<number[]>([1]);
   const [currentIndex, setCurrentIndex] = useState<number>(1);
-  const [pageQuestionCount, setPageQuestionCount] = useState<number>(8);
   const [questionAllCounts, setQuestionAllCounts] = useState<number>(40);
-
+  const { pageSize } = useResize();
   const handlecurrentIndex = (newValue: number) => {
     setCurrentIndex(newValue);
   };
@@ -27,13 +27,13 @@ const QuestionListPagination = ({ order }: QuestionListPaginationProps) => {
   }, [order]);
 
   useEffect(() => {
-    const maxIndex = Math.ceil(questionAllCounts / pageQuestionCount);
+    const maxIndex = Math.ceil(questionAllCounts / pageSize);
     setPageIndexes(
       Array(maxIndex)
         .fill(0)
         .map((_, i) => i + 1),
     );
-  }, [questionAllCounts, pageQuestionCount]);
+  }, [questionAllCounts, pageSize]);
 
   const handleQuestions = useCallback(
     async ({ limit, offset, sort }: SubjectListParams) => {
@@ -41,12 +41,12 @@ const QuestionListPagination = ({ order }: QuestionListPaginationProps) => {
       setQuestions(results);
       setQuestionAllCounts(count);
     },
-    [order, currentIndex],
+    [order, currentIndex, pageSize],
   );
 
   useEffect(() => {
     const sort = order === '이름순' ? 'name' : 'time';
-    handleQuestions({ limit: pageQuestionCount, offset: currentIndex, sort });
+    handleQuestions({ limit: pageSize, offset: currentIndex, sort });
   }, [handleQuestions]);
 
   return (
